+++ src/patch/dll_savestate.zig
@@ -34,11 +34,6 @@ const AxisInputMap = @import("core/Input.zig").AxisInputMap;
 // TODO: passthrough to annodue's panic via global function vtable; same for logging
 pub const panic = debug.annodue_panic;
 
-// FIXME: remove, for testing
-const dbg = @import("util/debug.zig");
-var gpa = std.heap.GeneralPurposeAllocator(.{}){};
-var alloc = gpa.allocator();
-
 // Usable in Practice Mode only
 
 // FEATURES
@@ -151,7 +146,7 @@ const state = struct {
     // FIXME: better new-frame checking that doesn't only account for tabbing out
     // i.e. also when pausing, physics frozen with ingame feature, etc.
     fn saveable(gs: *GlobalSt) bool {
-        return gs.in_race.on() and rec_data.saveable();
+        return gs.in_race.on() and rec_data.canSave();
     }
 
     // FIXME: check if you're actually in the racing part, also integrate with global
@@ -187,7 +182,7 @@ const state = struct {
 
 fn DoStateRecording(gs: *GlobalSt, _: *GlobalFn) LoadState {
     if (state.saveable(gs))
-        state.rec_data.save_compressed(gs.framecount);
+        state.rec_data.save(gs.framecount);
 
     if (state.save_input_st.gets() == .JustOn) {
         state.load_frame = state.rec_data.frame - 1;
@@ -208,7 +203,7 @@ fn DoStateLoading(gs: *GlobalSt, _: *GlobalFn) LoadState {
     }
     if (gs.timestamp >= state.load_time) {
         if (!state.loadable(gs)) return .Recording;
-        state.rec_data.load_compressed(state.load_frame);
+        state.rec_data.restore(state.load_frame);
         state.load_count += 1;
         return .Recording;
     }
@@ -232,13 +227,13 @@ fn DoStateScrubbing(gs: *GlobalSt, _: *GlobalFn) LoadState {
     );
 
     if (!state.loadable(gs)) return .Scrubbing;
-    state.rec_data.load_compressed(std.math.cast(u32, state.scrub_frame).?);
+    state.rec_data.restore(std.math.cast(u32, state.scrub_frame).?);
     return .Scrubbing;
 }
 
 fn DoStateScrubExiting(gs: *GlobalSt, _: *GlobalFn) LoadState {
     if (state.loadable(gs))
-        state.rec_data.load_compressed(std.math.cast(u32, state.scrub_frame).?);
+        state.rec_data.restore(std.math.cast(u32, state.scrub_frame).?);
 
     if (state.save_input_st.gets() == .JustOn) {
         state.load_frame = state.rec_data.frame - 1;
@@ -318,42 +313,6 @@ export fn InputUpdateB(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
 export fn EngineUpdateStage20A(gs: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
     if (!state.savestate_enable) return;
     UpdateState(gs, gf);
-
-    // FIXME: remove, for testing
-    //if (gf.InputGetKb(.J, .JustOn)) blk: {
-    //    const file = std.fs.cwd().createFile("annodue/test_frame_data_dump.bin", .{}) catch break :blk;
-    //    defer file.close();
-    //    var file_w = file.writer();
-    //    state.rec_data.write(file_w, .{ .frames = 256, .headers = false }) catch break :blk;
-    //    dbg.ConsoleOut(
-    //        "frame data written to file\n  frame size: 0x{X}\n",
-    //        .{state.rec_data.frame_size},
-    //    ) catch {};
-    //}
-    if (gf.InputGetKb(.F, .JustOn)) blk: {
-        const file = std.fs.cwd().createFile("annodue/recording_raw-0x2458-nohead_mgs.bin.txt", .{}) catch
-            break :blk;
-        defer file.close();
-        RewindDataType.calcPotential(alloc, file.writer(), .{
-            .compression = .bfid,
-            .sub_path = "annodue/recording_raw-0x2458-nohead_mgs.bin",
-            .frame_size = 0x2458,
-        }) catch {};
-        RewindDataType.calcPotential(alloc, file.writer(), .{
-            .compression = .xrle,
-            .sub_path = "annodue/recording_raw-0x2458-nohead_mgs.bin",
-            .frame_size = 0x2458,
-            .item_size = 4,
-            .rle_head_size = 2,
-        }) catch {};
-        RewindDataType.calcPotential(alloc, file.writer(), .{
-            .compression = .xrles,
-            .sub_path = "annodue/recording_raw-0x2458-nohead_mgs.bin",
-            .frame_size = 0x2458,
-            .item_size = 1,
-            .rle_head_size = 1,
-        }) catch {};
-    }
 }
 
 // TODO: merge with EngineUpdateStage20A?